rename activity -> notifications
